refactor: move arg parsing into new folder
