+++ types/arguments/index.d.ts
@@ -1,0 +1,1 @@
+export declare function parseArguments(rawArgs: any): any;
